Staging sites: rename is-status-reverting
